[AWIBOF-3429] Fix journal integration test bug
Block map list of StripeTestFixture which is for keeping stripe information
in testing is updated after the block logs are written. However,
when estimating remaining log buffer to decide logs for the stripe can be written,
it is required to update block map list before writing it.

So change to update block map list first before reuqest write test fixture to write it.


Former-commit-id: cc892067bfe43f0ab33b701f93e49c8bc8683e77