[AWIBOF-4094] Add field separator option for output
Add a global field separator option (--fs) for output.
When specified, each row of the output will be separate
with the separator.

Owing to this option, the CLI output can be pipelined to awk.
This enables functionalities such as conditional display of arrays,
devices, and volumes.

For example, you can display devices of which size is greater
than a value, or you can display unmounted arrays only.

Signed-off-by: Myoungjun Lee <jun20.lee@samsung.com>

Former-commit-id: 46bcf75d217bb6712fe8d5f14a07dc095792c2f4