[SHPOS-598] Delete every volumes in an array during array deletion
When an array is deleted, pos bdevs that belong to the array didn't get deleted.
So when array was created with the same name, pos bdev was reused,
and it caused segmentation fault in reactor.
So, fixed to delete all volume in the array and added init pos bdev during creation.

Signed-off-by: syeon.shin <syeon.shin@samsung.com>