[AWIBOF-5657] Fix command response duplication
Fix a bug where CLI server sends the response for the
previous CLI request.

This bug is because of CLI threads' race condition.
It occurs when a new CLI client connects to the CLI server
before the completion on the previous CLI client's request.

Previously, the CLI server just tested a lock using an unblocking
method, and used the preivous response string when it cannot acquire
the lock.

Change pthread_mutex_trylock() to pthread_mutex_lock to resolve this
iproblem.

Delete notifyDone check.

Signed-off-by: Myoungjun Lee <jun20.lee@samsung.com>