[AWIBOF-5685] Change CLI client output when POS is not running
Change CLI client output when POS is not running.

Previously, when POS is not running, the error message could not
be parsed as JSON because both socket error messages and json
output are displayed.

Now the CLI client builds and displays an error response (in JSON
form when --json-res flag is input) when it cannot connect to the
CLI server because POS is not running.

Add unit testing accordingly.

Signed-off-by: Myoungjun Lee <jun20.lee@samsung.com>