[SHPOS-666] bugfix for spdk - known issue
nvmf: check for null admin_qpair when updating subsystem pg
It is possible for a controller to get added to the subsystem before its admin_qpair has been assigned.
We need to account for that when traversing the subsystem's ctrlr list when determining ns and ana_changes that need to be reported for the ctrlr.

Found while doing stress testing with connects and subsystem ns add/remove.

https://github.com/spdk/spdk/commit/65ef1f32a63bfc43f8d67e939b551f6dcf976e54