[AWIBOF-6135] Change fields in Status protobuf message to optional
Change the fields in the Status protobuf message to optional
to let the code can be a zero value.

Without the optional flag, a zero value in the code field will
be recognized as no value by gRPC. This should not occur as the
result of the command becomes unknown.

Signed-off-by: mjlee34 <xelos2@naver.com>