[AWIBOF-6187] Add the "command" field to event log
Add the "command field to event log (both for structured and plain-
text log record).

Previously, it is hard to know which command PoseidonOS is processing
for each log record. This makes it difficult to understand the semantics
and contexts of the event logs, which leads to log analysis time.

For the first step to resolve the problem, this commit adds a field
, called "command" to each log record. "command" field indicates the
command in progress. This commit allows the developers to identify what
command triggers the events in the log.

The value of the command field is stored to a variable in Logger.
Add setter and resetter functions for the variable for future use.

Signed-off-by: mjlee34 <jun20.lee@samsung.com>