[AWIBOF-6206] Limit the number of concurrent clients to gRPC server to one
Limit the number of concurrent clients to gRPC server to one
(only one request will be served at a time).

Without this limit, it may happen that concorrent command
executions cause a crash of PoseidonOS. For example, it is not
allowed and supported to create multiple arrays at the same time.

This commit limits the number of concurrent clients to gRPC server,
by setting the reousrce quota of the gRPC server to 1.