[AWIBOF-6206] Add an ResourceExhausted handling to CLI client
Add an ResourceExhausted handling to CLI client.

If a CLI client sends a request to the CLI server when PoseidonOS
is processing a command, the CLI server must inform the CLI client
that it cannot process the request.

This commit adds a handling when the CLI sever sends a request with
ResourceExhausted status, which means that PoseidonOS is busy
to process another command.

The CLI client will tell the user to try the command after a while.

Limitation and to-do: this approach can guarantee that PoseidonOS does
not accept concurrent commands. However, it would be better if
the CLI server can process multiple commands in the order of arrivals.
I leave this as future work as it requires a request queue implementation.