[AWIBOF-6215] Change Telemetry commands to gRPC (#77)
* [AWIBOF-6206] Limit the number of concurrent clients to gRPC server to one

Limit the number of concurrent clients to gRPC server to one
(only one request will be served at a time).

Without this limit, it may happen that concorrent command
executions cause a crash of PoseidonOS. For example, it is not
allowed and supported to create multiple arrays at the same time.

This commit limits the number of concurrent clients to gRPC server,
by setting the reousrce quota of the gRPC server to 1.

* [AWIBOF-6206] Add an ResourceExhausted handling to CLI client

Add an ResourceExhausted handling to CLI client.

If a CLI client sends a request to the CLI server when PoseidonOS
is processing a command, the CLI server must inform the CLI client
that it cannot process the request.

This commit adds a handling when the CLI sever sends a request with
ResourceExhausted status, which means that PoseidonOS is busy
to process another command.

The CLI client will tell the user to try the command after a while.

Limitation and to-do: this approach can guarantee that PoseidonOS does
not accept concurrent commands. However, it would be better if
the CLI server can process multiple commands in the order of arrivals.
I leave this as future work as it requires a request queue implementation.

* [AWIBOF-6206] Refactor gRPC response status handling

Refactor gRPC response status handling.

Create a handler function for gRPC response error handling.
The handler switches by the status of a gRPC response, and
prints out a corresponding error message defined in strings.go

Make minor fixes to the codes for better readability.

* [AWIBOF-6215] Change Telemetry commands to gRPC

Change the communication protocol for Telemetry commands
to gRPC.

Add related events.

* [AWIBOF-6215] Fix a typo

Fix a typo in STOPTELEMETRY command.