[AWIBOF-6674] Count and report TELEMETRY_CLIENT_PUBLISHREQUEST_SEND_FAILURE event
The TELEMETRY_CLIENT_PUBLISHREQUEST_SEND_FAILURE event occurs when
PoseidonOS fails to send a gRPC message to exporter. The problem is that
the events are logged too much because as they are logged for every
gRPC message send.

Consequently, it is very hard for developers to see important log records
because of the TELEMETRY_CLIENT_PUBLISHREQUEST_SEND_FAILURE event.

This commit adds a count and a report threshold for logging the event.
PoseidonOS records the log event only when the event occurs more than a threshold.

Signed-off-by: mjlee34 <jun20.lee@samsung.com>