[AWIBOF-6740] Add a node name to CLI as a global flag
Add a node name (for high-availbaility) to CLI as a global flag.

When the users want to send a CLI command to a specific node,
this commit allows them to specify the name of the node using the --node flag.

When the CLI client executes the command, it fetches the IP address to the node
by looking up the database having the information about the node.

Signed-off-by: mjlee34 <jun20.lee@samsung.com>