[AWIBOF-6879] Set timeout to 30 min for UNMOUNTARRAY command
Set timeout to 30 min for UNMOUNTARRAY command.

Failures in CI tests are often because
UNMOUNTARRAY command exceptionally takes long time.

To workaround this problem, this commit sets the timeout for the command
to 30 minutes.

However, the timeout needs to be set based on a systemical approach
rather than a temporal constant.

Signed-off-by: mjlee34 <jun20.lee@samsung.com>