[AWIBOF-6773] Copy pos_event.yaml for unit testing
Copy pos_event.yaml to the pos_conf directory for unit testing.

Previously, unit testing failed because pos_event.yaml was not copied
to the config directory. It was because unit testing script
did not execute $make install, which copies pos_event.yaml to the config
directory.

To fix the problem, this commit adds a copy command to the unit testing
script for github action.

Signed-off-by: mjlee34 <jun20.lee@samsung.com>