[AWIBOF-7397] Add a method to display a protobuf CLI request in json
Add a method to display a protobuf CLI request in json.

Previously, the CLI client converts a CLI request in protobuf form
to json form when specifying --json-req flag. The problem was
that the readability was poor because the conversion was done
in the high-level function that interpretes a command line.

This commit adds PrintProtoReqJson(req protoreflect.ProtoMessage)
that converts a protobuf request to json and display the request to stdout.

Signed-off-by: mjlee34 <jun20.lee@samsung.com>