[AWIBOF-7399] Add BurstFilter to logger
Add BurstFilter to logger.

Previously, although a number of the same events occur,
the old logger logged all the duplicated events.
This was problematic because the huge number of event logs
severely degraded the readability of the log.

To mitigate the problem, this commit implements BurstFilter,
which prevents event log duplication.

When the same events occur repeatedly, the new logger does not
log the event. Instead, it stores the event's ID, message,
and the repetition count. The duplicated events are logged to the file
when a new (different) event occurs.

Signed-off-by: mjlee34 <jun20.lee@samsung.com>