[AWIBOF-7500] Increase CLI timeout for MOUNTARRAY to 10 min.
Increase CLI timeout for MOUNTARRAY command to 10 minutes.

Previously, the MOUNTARRAY command was terminated by a CLI timeout
because its execution time is longer than the default timeout value.

To prevent this undesirable timeout, this commit increases the timeout
value for MOUNTARRAY command to 10 minutes.

Note: this approach is a workaround; the fundamental
solution for this problem is to apply a proper timeout value for each
command. To decide the proper timeout value, execution time analysis
is required.

Signed-off-by: mjlee34 <jun20.lee@samsung.com>