[AWIBOF-8171] Fix subsystem existance checking logic during volume mount (#763)
* [AWIBOF-8171] Fix subsystem existance checking logic during volume mount

Subsystem belongs to an array at the point of volume mount. Before
mounting volume, it first checks if the requested subsystem exists
or any subsystem exists to mount. However, the logic for checking
subsystem didn't consider relation of subsystem with array.
It should check either the requested subsystem belongs to the same
array with volume or not belong to any array.

Signed-off-by: syeon.shin <syeon.shin@samsung.com>