[AWIBOF-8218] Change SegmentFreeUpdateRequest event type to callback
Why:
- Improve debuggability by changing event type to callback

What:
- Change segment free update request event type from 'event' to 'callback'

This commit changes the type of SegmentFreeUpdateRequest from 'event' to 'callback'.
The reason for this change is that when using the 'event' type, it can be difficult to trace events during debugging.
By changing the event type to 'callback', we can improve debuggability and make it easier to trace events during development and testing.

Signed-off-by: cheolho.kang <cheolho.kang@samsung.com>